requirement.txt
altgraph==0.17.3
autopep8==2.0.2
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
kiwisolver==1.4.4
macholib==1.16.2
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pycodestyle==2.10.0
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyparsing==3.0.9
python-dateutil==2.8.2
reportlab==3.6.12
six==1.16.0
